requirements.txt
torchvision==0.11.2
tensorboard
wandb
torchinfo